pyproject.toml
[project]
name = "constitutional-reliability-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "faiss-cpu",
    "fastapi>=0.128.8",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langgraph>=0.6.11",
    "lxml>=6.0.2",
    "pydantic>=2.12.5",
    "pypdf>=6.7.4",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "uvicorn>=0.39.0",
]
